Add files via upload
This is my first version of the webpage I tried to create. I added different heading, colours, picture and also links to my website. I also find out that those https links are not working and I would really appreciate if I would be given an advice how to do it correctly.